Revert "Merge branch 'feature/modal-dialog' of https://github.com/mash-up-kr/mash-up-recruit-fe into env/storybook-preview"
This reverts commit 661fa227b6921f5026b7bf4bfa3ffc8fb5beda37, reversing
changes made to e037c1235f86a47156fd0f29668e923922625311.
