+++ emotion.d.ts
@@ -1,10 +1,9 @@
 import '@emotion/react';
-import { FontType, ColorType, ZIndexType } from '@/styles';
+import { FontType, ColorType } from '@/styles';
 
 declare module '@emotion/react' {
   export interface Theme {
     fonts: FontType;
     colors: ColorType;
-    zIndex: ZIndexType;
   }
 }
