+++ package.json
@@ -36,7 +36,6 @@
     "@svgr/webpack": "^6.2.1",
     "@types/node": "17.0.12",
     "@types/react": "17.0.38",
-    "@types/react-dom": "^17.0.11",
     "babel-loader": "^8.2.3",
     "chromatic": "^6.4.3",
     "eslint": "8.7.0",
