+++ pages/_document.tsx
@@ -49,7 +49,6 @@ class MyDocument extends Document {
         </Head>
         <body>
           <Main />
-          <div id="modal-root" />
           <NextScript />
         </body>
       </Html>
