+++ src/components/common/index.ts
@@ -1,9 +1,4 @@
-export { default as LinkTo } from './LinkTo/LinkTo.component';
-export { default as MainNavigation } from './MainNavigation/MainNavigation.component';
-export { default as Header } from './Header/Header.component';
-export { default as Layout } from './Layout/Layout.component';
-export { default as Portal } from './Portal/Portal.component';
-export { default as Modal } from './Modal/Modal.component';
-export { default as LoginModalDialog } from './LoginModalDialog/LoginModalDialog.component';
-export { default as AlertModalDialog } from './AlertModalDialog/AlertModalDialog.component';
-export { default as ConfirmModalDialog } from './ConfirmModalDialog/ConfirmModalDialog.component';
+export { default as LinkTo } from './LinkTo/LinkTo';
+export { default as MainNavigation } from './MainNavigation/MainNavigation';
+export { default as Header } from './Header/Header';
+export { default as Layout } from './Layout/Layout';
