+++ src/styles/themes/index.ts
@@ -1,4 +1,3 @@
 export * from './colors';
 export * from './fonts';
-export * from './zIndex';
 export * from './theme';
