+++ src/styles/themes/theme.ts
@@ -1,9 +1,7 @@
 import { fonts } from './fonts';
 import { colors } from './colors';
-import { zIndex } from './zIndex';
 
 export const theme = {
   fonts,
   colors,
-  zIndex,
 };
