--- src/styles/themes/zIndex.ts
@@ -1,7 +1,0 @@
-export const zIndex = {
-  default: 0,
-  modal: 5,
-  dialog: 10,
-};
-
-export type ZIndexType = typeof zIndex;
